requirements.txt
-i https://pypi.org/simple
anyio==3.6.2
arrow==1.2.3
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
et-xmlfile==1.1.0 ; python_version >= '3.6'
greenlet==2.0.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
h11==0.14.0 ; python_version >= '3.7'
hashids==1.3.1
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
idna==3.4 ; python_version >= '3.5'
lxml==4.9.2
markdown==3.4.1
numpy==1.24.1 ; python_version >= '3.8'
nvidia-cublas-cu11==11.10.3.66 ; platform_system == 'Linux'
nvidia-cuda-nvrtc-cu11==11.7.99 ; platform_system == 'Linux'
nvidia-cuda-runtime-cu11==11.7.99 ; platform_system == 'Linux'
nvidia-cudnn-cu11==8.5.0.96 ; platform_system == 'Linux'
openai==0.25.0
openpyxl==3.0.10 ; python_version >= '3.6'
pandas==1.5.3 ; python_version >= '3.8'
pandas-stubs==1.5.2.230105 ; python_version < '3.12' and python_version >= '3.8'
pillow==9.4.0
playwright==1.29.1
pony==0.7.16
psutil==5.9.4
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.4.8
pydantic[dotenv]==1.10.4
pyee==9.0.4
pyrogram==2.0.97
pysocks==1.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rfc3986[idna2008]==1.5.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
setuptools==66.1.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
telethon==1.26.1
tgcrypto==1.2.5
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transliterate==1.10.2
types-pytz==2022.7.1.0
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14
wheel==0.38.4 ; python_version >= '3.7'
